Add async ParameterTree implementation + refactored base class
